+ Since numpy_1.24 types like np.int, np.float and np.bool are no longer used, they have been substituted Python native types int, float and bool respectively.     + Dockefile with container for Ubuntu 20.04 was added by DimitriosBellos

+++ code/pyorg/scripts/mt_lumen/particles_curator.py
@@ -111,7 +111,7 @@ except pexceptions.PySegInputError as e:
     sys.exit(-1)
 part_dsts, part_end = dict().fromkeys(list(range(star.get_nrows()))), dict().fromkeys(list(range(star.get_nrows())))
 for key in part_dsts.keys():
-    part_dsts[key] = np.finfo(np.float).max
+    part_dsts[key] = np.finfo(float).max
     part_end[key] = np.asarray((0., 0., 0.), dtype=float)
 
 if p_bin > 0:
